% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO-GN.R
\docType{data}
\name{CO_GN}
\alias{CO_GN}
\title{Condylion-Gnathion}
\format{
A data.frame with 22 observations of the following 10 variables: \describe{
  \item{Variable_Num}{Variable number of Riolo}
  \item{Type}{Variable type (Angular or Linear)}
  \item{Long_Name}{Descriptive variable name}
  \item{Short_Name}{Abbreviated variable name}
  \item{Numeric_Name}{Numeric landmark coding}
  \item{Age}{Age in years}
  \item{Sex}{Female or Male}
  \item{N}{Numeber of observations}
  \item{Mean}{Mean value}
  \item{SD}{Standard deviation}
}
}
\description{
Variable CO-GN from Riolo et al. (1974)
}
\keyword{datasets}
