% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Riolo_data.R
\name{get_Riolo_data}
\alias{get_Riolo_data}
\title{Load Riolo Data}
\usage{
get_Riolo_data(Riolo_var, sex, rescale = FALSE)
}
\arguments{
\item{Riolo_var}{string Riolo variable name}

\item{sex}{string for sex}

\item{rescale}{Boolean for whether to rescale data using Michigan scaling
factor}
}
\value{
data.frame with Riolo data
}
\description{
Load Riolo Data
}
