% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_Riolo.R
\name{match_Riolo}
\alias{match_Riolo}
\title{Match Riolo Data}
\usage{
match_Riolo(tr)
}
\arguments{
\item{tr}{string of trait to match}
}
\value{
string Riolo variable name
}
\description{
Match Riolo Data
}
