% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Riolo.R
\docType{package}
\name{Riolo-package}
\alias{Riolo-package}
\alias{Riolo}
\title{Data sets from Riolo et al. (1974)}
\description{
The \code{Riolo} package contains data sets from Riolo ML, Moyers RE,
  McNamara JA, Hunter WS. 1974. An atlas of
  craniofacial growth: cephalometric standards from the University School
  Growth Study, the University of Michigan. Craniofacial Growth Series
  Monograph 2.
}
\details{
\tabular{ll}{ Package: \tab Riolo\cr
              Type: \tab Package\cr
              Version: \tab 0.1.0\cr
              Date: \tab 2020-07-28\cr
              License: \tab MIT\cr
              LazyLoad: \tab yes\cr
              LazyData: \tab yes\cr }
}
\references{
Riolo ML, Moyers RE, McNamara JA, Hunter WS. 1974. An atlas of
  craniofacial growth: cephalometric standards from the University School
  Growth Study, the University of Michigan. Craniofacial Growth Series
  Monograph 2.
}
\author{
Kevin M. Middleton (\url{middletonk@missouri.edu})
}
\keyword{package}
